Refactor reply parsing
