+++ cosmwasm/enclaves/shared/contract-engine/src/lib.rs
@@ -18,6 +18,7 @@ mod io;
 mod message;
 mod query_chain;
 mod random;
+pub mod reply_message;
 pub(crate) mod types;
 #[cfg(feature = "wasmi-engine")]
 mod wasm;
